refactor: move artifact specifier parsing to pkg/util
